dvc pull added as per md file

+++ inference_api.py
@@ -39,9 +39,9 @@ class RowData(BaseModel):
 
 if "DYNO" in os.environ and os.path.isdir(".dvc"):
     os.system("dvc config core.no_scm true")
-    os.system("dvc pull")
-    # if os.system("dvc pull") != 0:
-        # exit("dvc pull failed")
+    # os.system("dvc pull")
+    if os.system("dvc pull") != 0:
+        exit("dvc pull failed")
     os.system("rm -r .dvc .apt/usr/lib/dvc")
